Floating fees for the client
Estimate what fee or priority is needed to get a transaction
accepted into a block (unless user has set a fee using
-paytxfee or the GUI).

Includes a safety net: always pay at least what a 0.8.5 client
would pay. The plan is for future releases to eliminate that
check, and let fees float up and down based on competition for
block space.

Adds two new RPC methods to expose the estimates:
estimatefee and estimatepriority.

I expect future releases might improve the fee estimation code;
it is self-contained in the CMinerPolicyEstimator class, so
"patches welcome."

This is a much more conservative of the pull request I
submitted a couple months ago.